requirements.txt
appnope==0.1.0
attrs==19.3.0
backports-abc==0.5
backports.csv==1.0.7
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.8.2
bleach==3.1.1
blis==0.4.1
boto==2.49.0
boto3==1.12.12
botocore==1.15.12
bz2file==0.98
cachetools==3.1.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
docutils==0.15.2
entrypoints==0.3
enum34==1.1.9
feedparser==5.2.1
functools32==3.2.3.post2
future==0.18.2
futures==3.3.0
gensim==3.8.1
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-language==1.3.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
importlib-metadata==1.5.0
ipaddress==1.0.23
ipykernel==4.10.1
ipython==5.9.0
ipython-genutils==0.2.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.3
jupyterlab==0.33.12
jupyterlab-launcher==0.11.2
kiwisolver==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==2.2.5
mistune==0.8.4
murmurhash==1.0.2
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==4.4.0
nltk==3.4.5
notebook==5.7.8
numpy==1.16.1
pandas==0.24.2
pandocfilters==1.4.2
pathlib==1.0.1
pathlib2==2.3.5
patsy==0.5.1
Pattern==2.6
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
preshed==3.0.2
progressbar2==3.50.1
prometheus-client==0.7.1
prompt-toolkit==1.0.18
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
pygam==0.8.0
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
-e git+https://github.com/sabjoslo/pyspan.git@1ba12e367e6d782fc7515ef3c787503e07065d89#egg=pyspan
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2019.3
pyzmq==19.0.0
requests==2.23.0
rpy2==2.8.6
rsa==4.0
s3transfer==0.3.3
scandir==1.10.0
scikit-learn==0.20.4
scipy==1.2.3
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
smart-open==1.9.0
soupsieve==1.9.5
spacy==2.2.4
srsly==1.0.2
statsmodels==0.10.2
subprocess32==3.5.4
tabletext==0.1
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
tornado==5.1.1
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
-e git+https://github.com/sabjoslo/wordplay.git@a1855f399e71b5854d0d8925a58103ec9272e9ec#egg=wordplay
zipp==1.2.0
